refactor: use package extensions and reduce package dependencies
Since AtomsIO is meant to be a standardized package that allows us to
read/write atomic structures, it makes natural sense that all the
different packages it supports are made package extensions and aren't
direct dependencies unnecessarily making the package bigger than it
needs to be.
