+++ Project.toml
@@ -3,16 +3,23 @@ uuid = "1692102d-eeb4-4df9-807b-c9517f998d44"
 authors = ["Michael F. Herbst <info@michael-herbst.com> and contributors"]
 version = "0.2.3"
 
-[deps]
-AtomsBase = "a963bdd2-2df7-4f54-a1ee-49d51e6be12a"
+[weakdeps]
 Chemfiles = "46823bd8-5fb3-5f92-9aa0-96921f3dd015"
 ExtXYZ = "352459e4-ddd7-4360-8937-99dcb397b478"
+XCrySDenStructureFormat = "02310045-4665-40c9-a658-98c497d2eca9"
+
+[extensions]
+AtomsIOChemfilesExt = "Chemfiles"
+AtomsIOExtXYZExt = "ExtXYZ"
+AtomsIOXCrySDenStructureFormatExt = "XCrySDenStructureFormat"
+
+[deps]
+AtomsBase = "a963bdd2-2df7-4f54-a1ee-49d51e6be12a"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 PeriodicTable = "7b2266bf-644c-5ea3-82d8-af4bbd25a884"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 UnitfulAtomic = "a7773ee8-282e-5fa2-be4e-bd808c38a91a"
-XCrySDenStructureFormat = "02310045-4665-40c9-a658-98c497d2eca9"
 
 [compat]
 AtomsBase = "0.3"
@@ -33,4 +40,4 @@ Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["AtomsBaseTesting", "LinearAlgebra", "Pkg", "Test"]
+test = ["AtomsBaseTesting", "LinearAlgebra", "Pkg", "Test", "Chemfiles", "ExtXYZ", "XCrySDenStructureFormat"]
