+++ ext/AtomsIOExtXYZExt.jl
@@ -1,14 +1,9 @@
+module AtomsIOExtXYZExt
+
 import ExtXYZ
+using AtomsIO
 
-"""
-Parse or write file using [ExtXYZ](https://github.com/libAtoms/ExtXYZ.jl)
-
-Supported formats:  
-  - [XYZ](https://openbabel.org/wiki/XYZ) and [extxyz](https://github.com/libAtoms/extxyz#extended-xyz-specification-and-parsing-tools) files
-"""
-struct ExtxyzParser <: AbstractParser end
-
-function supports_parsing(::ExtxyzParser, file; save, trajectory)
+function AtomsIO.supports_parsing(::ExtxyzParser, file; save, trajectory)
     _, ext = splitext(file)
     ext in (".xyz", ".extxyz")
 end
@@ -34,7 +29,7 @@ function _extxyz_read_frames(args...; kwargs...)
     frames
 end
 
-function load_system(::ExtxyzParser, file::AbstractString, index=nothing)
+function AtomsIO.load_system(::ExtxyzParser, file::AbstractString, index=nothing)
     if isnothing(index)
         frames = _extxyz_read_frames(file)
         return ExtXYZ.Atoms(last(frames))
@@ -44,15 +39,17 @@ function load_system(::ExtxyzParser, file::AbstractString, index=nothing)
     end
 end
 
-function save_system(::ExtxyzParser, file::AbstractString, system::AbstractSystem)
+function AtomsIO.save_system(::ExtxyzParser, file::AbstractString, system::AbstractSystem)
     ExtXYZ.write_frame(file, ExtXYZ.write_dict(system))
 end
 
-function load_trajectory(::ExtxyzParser, file::AbstractString)
+function AtomsIO.load_trajectory(::ExtxyzParser, file::AbstractString)
     ExtXYZ.Atoms.(_extxyz_read_frames(file))
 end
 
-function save_trajectory(::ExtxyzParser, file::AbstractString,
+function AtomsIO.save_trajectory(::ExtxyzParser, file::AbstractString,
                          systems::AbstractVector{<:AbstractSystem})
     ExtXYZ.write_frames(file, ExtXYZ.write_dict.(systems))
+end
+
 end
