+++ src/AtomsIO.jl
@@ -4,12 +4,9 @@ using Reexport
 import PeriodicTable
 
 include("parser.jl")
-include("chemfiles.jl")
-include("extxyz.jl")
-include("xsf.jl")
-include("saveload.jl")
+export AbstractParser, ChemfilesParser, ExtxyzParser, XcrysdenstructureformatParser
 
+include("saveload.jl")
 export load_system, save_system, load_trajectory, save_trajectory
-export AbstractParser
-export ChemfilesParser, ExtxyzParser, XcrysdenstructureformatParser
+
 end
